requirements.txt
PyQt5
PyQt5-tools
Pillow
PyPDF2
selenium
beautifulsoup4
BeautifulSoup
re
json
requests
